requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==2.1.5
paho-mqtt==2.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4

pytest~=8.3.3